+++ modules/3d/src/assets/prefab_loader.rs
@@ -1,89 +1,76 @@
-use std::io::Read;
+use std::io::Cursor;
 use std::sync::Arc;
 
+use crayon::application::Context;
 use crayon::bincode;
 use crayon::errors::*;
-use crayon::res::location::Location;
-use crayon::res::{ResourceHandle, ResourceLoader, ResourceSystemShared};
+use crayon::res::registry::Register;
+use crayon::res::ResourceSystemShared;
+use crayon::video::VideoSystemShared;
 
 use super::prefab::*;
-use super::WorldResourcesShared;
 
 pub const MAGIC: [u8; 8] = [
     'P' as u8, 'R' as u8, 'E' as u8, 'B' as u8, ' ' as u8, 0, 0, 1,
 ];
 
+#[derive(Clone)]
 pub struct PrefabLoader {
-    world_resources: Arc<WorldResourcesShared>,
+    video: Arc<VideoSystemShared>,
     res: Arc<ResourceSystemShared>,
 }
 
 impl PrefabLoader {
-    pub fn new(res: Arc<ResourceSystemShared>, world_resources: Arc<WorldResourcesShared>) -> Self {
+    pub fn new(ctx: &Context) -> Self {
         PrefabLoader {
-            res: res,
-            world_resources: world_resources,
+            res: ctx.res.clone(),
+            video: ctx.video.clone(),
         }
     }
 }
 
-impl ResourceHandle for PrefabHandle {
-    type Loader = PrefabLoader;
-}
-
-impl ResourceLoader for PrefabLoader {
+impl Register for PrefabLoader {
     type Handle = PrefabHandle;
+    type Intermediate = Prefab;
+    type Value = Arc<Prefab>;
 
-    fn create(&self) -> Result<Self::Handle> {
-        let handle = self.world_resources.create_prefab_async();
-        info!("[PrefabLoader] creates {:?}.", handle);
-        Ok(handle)
-    }
-
-    fn load(&self, handle: Self::Handle, mut file: &mut dyn Read) -> Result<()> {
-        let mut buf = [0; 8];
-        file.read_exact(&mut buf[0..8])?;
-
-        // magic: [u8; 8]
-        if &buf[0..8] != &MAGIC[..] {
+    fn load(&self, handle: Self::Handle, bytes: &[u8]) -> Result<Self::Intermediate> {
+        if &bytes[0..8] != &MAGIC[..] {
             bail!("[PrefabLoader] MAGIC number not match.");
         }
 
-        let mut data: Prefab = bincode::deserialize_from(&mut file)?;
-        for v in &data.universe_meshes {
-            data.meshes.push(self.res.load_from(Location::from(*v))?);
+        let mut file = Cursor::new(&bytes[8..]);
+        let mut prefab: Prefab = bincode::deserialize_from(&mut file)?;
+
+        for &v in &prefab.universe_meshes {
+            let mesh = self.video.create_mesh_from_uuid(v)?;
+            prefab.meshes.push(mesh);
         }
 
-        for &v in &data.meshes {
-            self.res.wait(v)?;
-        }
+        // for &v in &prefab.meshes {
+        //     self.res.wait(v)?;
+        // }
 
         info!(
             "[PrefabLoader] loads {:?}. (Nodes: {}, Meshes: {})",
             handle,
-            data.nodes.len(),
-            data.meshes.len()
+            prefab.nodes.len(),
+            prefab.meshes.len()
         );
 
-        // The prefab handle might already been freed.
-        if let Some(prefab) = self.world_resources.update_prefab_async(handle, data)? {
-            for v in prefab.meshes {
-                self.res.unload(v)?;
-            }
-        }
-
-        Ok(())
+        Ok(prefab)
     }
 
-    fn delete(&self, handle: Self::Handle) -> Result<()> {
-        info!("[PrefabLoader] deletes {:?}.", handle);
+    fn attach(&self, handle: Self::Handle, item: Self::Intermediate) -> Result<Self::Value> {
+        info!("[PrefabLoader] attach {:?}.", handle);
+        Ok(Arc::new(item))
+    }
 
-        if let Some(prefab) = self.world_resources.delete_prefab_async(handle) {
-            for &v in &prefab.meshes {
-                self.res.unload(v)?;
-            }
+    fn detach(&self, handle: Self::Handle, prefab: Self::Value) {
+        info!("[PrefabLoader] detach {:?}.", handle);
+
+        for &v in &prefab.meshes {
+            self.video.delete_mesh(v);
         }
-
-        Ok(())
     }
 }